requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.7.8
openai==1.3.7
customtkinter==5.2.0
streamlit==1.28.1
jsonschema==4.20.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pytest==7.4.3
pytest-cov==6.2.1
cryptography==45.0.5
black==23.11.0
flake8==6.1.0 